requirements.txt
pygame
opencv-python
mediapipe
google-generativeai
python-dotenv
